admin panel categories add

+++ src/app/layout/products/products-list/products-list.component.ts
@@ -1,9 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, ViewChild} from '@angular/core';
 import { routerTransition } from '../../../router.animations';
 import { ApiService } from '../../../shared/services/api.service';
-import { MatTableDataSource } from '@angular/material';
+import { MatPaginator, MatTableDataSource } from '@angular/material';
 import { MatDialog, MatDialogRef, MAT_DIALOG_DATA } from '@angular/material';
 import { UpdateProductComponent } from '../update-product/update-product.component';
+import { ConfirmAskComponent } from '../../../shared/confirm-ask/confirm-ask.component';
 
 @Component({
   selector: 'app-products-list',
@@ -13,12 +14,13 @@ import { UpdateProductComponent } from '../update-product/update-product.compone
 
 })
 export class ProductsListComponent implements OnInit {
+  @ViewChild(MatPaginator) paginator: MatPaginator;
 
   departmentsList = [];
   subDepartmentsList = [];
   productList = [];
-  departmentId = 0;
-  subDepartmentId = 0;
+  departmentId = 1;
+  subDepartmentId = 1;
   displayedColumns: string[] = [
     'productId',
     'productName',
@@ -45,6 +47,8 @@ export class ProductsListComponent implements OnInit {
         this.departmentsList = response.data;
       }
     });
+    this.onDepartmentChange();
+    this.getProductList();
   }
 
   onDepartmentChange() {
@@ -55,9 +59,13 @@ export class ProductsListComponent implements OnInit {
 
   applyFilter(filterValue: string) {
     this.data.filter = filterValue.trim().toLowerCase();
+    if (this.data.paginator) {
+      this.data.paginator.firstPage();
+    }
   }
 
   getProductList() {
+    console.log('Product list called');
     const data = {
       departmentId: this.departmentId,
       subDepartmentId: this.subDepartmentId
@@ -68,6 +76,7 @@ export class ProductsListComponent implements OnInit {
       this.data = new MatTableDataSource(response.data);
       this.isLoadingResults = false;
       this.resultsLength = response.data.length;
+      this.data.paginator = this.paginator;
     });
   }
 
@@ -75,7 +84,7 @@ export class ProductsListComponent implements OnInit {
     console.log('print product data');
     console.log(prodData);
 
-   let setData = {
+    let setData = {
       accessToken: 'dEStSa4Jd52b',
       productId: prodData.productId,
       departmentId: 0,
@@ -83,6 +92,8 @@ export class ProductsListComponent implements OnInit {
       subCategoryId: 0,
       productName: prodData.productName,
       brandId: 0,
+      unitId: 0,
+      unit: prodData.unit,
       description: prodData.description,
       regularPrice: prodData.regularPrice,
       salePrice: prodData.salePrice,
@@ -113,6 +124,7 @@ export class ProductsListComponent implements OnInit {
         setData.subDepartmentId = response.data[0].sub_department_id !== 0 ? response.data[0].sub_department_id : '';
         setData.subCategoryId = response.data[0].sub_category_id !== 0 ? response.data[0].sub_category_id : '';
         setData.brandId = response.data[0].brand_id !== 0 ? response.data[0].brand_id : '';
+        setData.unitId = response.data[0].unit_id !== 0 ? response.data[0].unit_id : '';
         setData.taxStatusId = response.data[0].tax_status_id !== 0 ? response.data[0].tax_status_id : '';
         setData.taxClassId = response.data[0].tax_class_id !== 0 ? response.data[0].tax_class_id : '';
         setData.stockStatusId = response.data[0].stock_status_id !== 0 ? response.data[0].stock_status_id : '';
@@ -126,6 +138,7 @@ export class ProductsListComponent implements OnInit {
             data: setData,
           });
           dialogRef.afterClosed().subscribe(result => {
+            this.getProductList();
             console.log('Dialog closed');
           });
         });
@@ -137,9 +150,21 @@ export class ProductsListComponent implements OnInit {
       productId: id,
       accessToken: 'dEStSa4Jd52b'
     };
-    this.apiService.deleteProduct(deleteData).subscribe((response: any) => {
-      console.log(response);
+    const dialogRef = this.dialog.open(ConfirmAskComponent, {
+      width: '40%',
+      data: 'Are you sure you want to delete this product'
     });
-
+    dialogRef.afterClosed().subscribe(result => {
+      console.log(result);
+      console.log('Dialog closed');
+      if (result === true) {
+        this.apiService.deleteProduct(deleteData).subscribe((response: any) => {
+          console.log(response);
+          if (response.status === 200) {
+            this.getProductList();
+          }
+        });
+      }
+    });
   }
 }
